adds AddressModel and ActivityModel
