+++ src/store/modules/models/CongDefaultOptions.js
@@ -1,0 +1,30 @@
+export const CongDefault = {
+  options: {
+    territories: {
+      defaultSort: {
+        options: [
+          'Description',
+          'Name',
+        ],
+      },
+    },
+    territory: {
+      defaultView: {
+        options: [
+          { value: 'address-list', text: 'Address List' },
+          { value: 'phone-list', text: 'Phone List' },
+        ],
+      },
+    },
+    level1: {
+      level2: {
+        level3: {
+          options: [
+            'test1',
+            'test2',
+          ],
+        },
+      },
+    },
+  },
+};
